requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
idna==3.10
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
httpx==0.27.0
responses==0.25.7
python-dotenv==1.0.1
